% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteMartin.R
\name{paletteMartin}
\alias{paletteMartin}
\title{Discrete palette generator}
\usage{
paletteMartin(n = 15)
}
\arguments{
\item{n}{Number of colours to return. Max = 15.}
}
\value{
Returns a vector of colours of length n.
}
\description{
Generate a palette of up to 15 colours. The colours are from the paletteMartin palette in the colorBlindess R package.
}
\examples{
scales::show_col(paletteMartin(n = 9))
}
