% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_exp}
\alias{toy_exp}
\title{Toy expression data for CCPlotR}
\format{
\subsection{\code{toy_exp}}{

A data frame with 477 rows and 3 columns:
\describe{
\item{cell_type}{Cell type}
\item{gene}{Ligand/receptor gene}
\item{mean_exp}{Mean (normalised) expression of lignad/receptor gene in cell type}
}
}
}
\usage{
toy_exp
}
\description{
A dataframe showing the mean expression values for each ligand and receptor in each cell type.
}
\keyword{datasets}
