% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_network.R
\name{cc_network}
\alias{cc_network}
\title{Network Plot Function}
\usage{
cc_network(cc_df, colours = paletteMartin())
}
\arguments{
\item{cc_df}{A dataframe with columns 'source', 'target', 'ligand', 'receptor' and 'score'. See \code{toy_data} for example.}

\item{colours}{A vector of colours for each cell type. Default is \code{paletteMartin()}, a colourblind-friendly palette.}
}
\description{
This function plots a network of representing the number of interactions between cell types
}
\examples{
cc_network(toy_data)
cc_network(toy_data, colours = c('hotpink', 'orange', 'cornflowerblue'))
}
