% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy data for CCPlotR}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 735 rows and 5 columns.
}
\source{
This is a modified version of the toy dataset that comes with the Liana R package.
}
\usage{
data(toy_data)
}
\value{
\subsection{\code{toy_data}}{

A data frame with 735 rows and 5 columns:
\describe{
\item{source}{Cell type expressing the ligand}
\item{target}{Cell type expressing the receptor}
\item{ligand}{Ligand}
\item{receptor}{Receptor}
\item{score}{A score for each interaction e.g. -log10(aggregate_rank) returned by Liana}
}
}
}
\description{
A toy dataset of ligand-receptor interactions to demonstrate cell-cell interaction plots.
}
\keyword{datasets}
